Reorganizing instrumentation configuration
